+++ projects/ngz-material/src/lib/pipes/value/index.ts
@@ -1,0 +1,5 @@
+// Value processing pipes repository
+// ----------------------------------------------------------------------------
+
+// Import and (re)export all value processing pipes
+export * from './AsDate';